[CHK-905] fix: update unit tests
